test: Processor 테스트 분리
